test: Extract the tests for TestDMGMPermissionChecksForReactions into a separate file to avoid conflict
