+++ server/channels/app/post_test.go
@@ -4112,5 +4112,3 @@ func TestPopulateEditHistoryFileMetadata(t *testing.T) {
 		require.Greater(t, post2.Metadata.Files[0].DeleteAt, int64(0))
 	})
 }
-
-
